third_party/WebKit/LayoutTests/editing/deleting/delete-and-undo.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => {
    assert_selection(
        '<div contenteditable>foo ^bar |baz</div>',
        selection => {
            selection.document.execCommand('delete');
            selection.document.execCommand('undo');
        },
        '<div contenteditable>foo ^bar |baz</div>');
}, '"delete" then "undo" changes nothing');
</script>
